+++ users.go
@@ -9,7 +9,7 @@ import (
 type User struct {
 	Name                 *string             `json:"name,omitempty"`
 	Email                *string             `json:"email,omitempty"`
-	UID                  *int                `json:"uid,omitempty"`
+	UID                  *int64              `json:"uid,omitempty"`
 	LastLogin            *time.Time          `json:"lastLogin,omitempty"`
 	DateRegistered       *time.Time          `json:"dateRegistered,omitempty"`
 	LoginAccountGroup    *AccountGroup       `json:"loginAccountGroup,omitempty"`
